refactor: #76 서버 유틸 추가해서 중복 제거
